src/main/java/com/example/warehousemanagement/service/CategoryService.java
package com.example.warehousemanagement.service;

import com.example.warehousemanagement.model.Category;
import com.example.warehousemanagement.repository.CategoryRepository;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class CategoryService {
    private final CategoryRepository categoryRepository;

    public CategoryService(CategoryRepository categoryRepository) {
        this.categoryRepository = categoryRepository;
    }

    public List<Category> findAll() {
        return categoryRepository.findAll();
    }

    public Category save(Category category) {
        return categoryRepository.save(category);
    }

    public void delete(Long id) {
        categoryRepository.deleteById(id);
    }

    public Optional<Category> findById(Long id) {
        return categoryRepository.findById(id);
    }

    public Optional<Category> getCategoryById(long l) {
        return categoryRepository.findById(l);
    }

    public Category createCategory(Category category) {return categoryRepository.save(category);
    }
}